Add quota_factor for maas_rally scenarios
This change adds a quota_factor variable for each scenario.  This value is used
to calculate an effective quota for each scenario's project.  The primary use
case is to provide enough buffer in the effective quotas to allow a new
invocation of a scenario to run while a failed invocation is being cleaned up.